+++ CASWebApi/Controllers/HolidayController.cs
@@ -33,10 +33,15 @@ namespace CASWebApi.Controllers
             logger.LogInformation("Getting all Holidays from HolidayController");
             var holidayList = _holidayService.GetAll();
             if (holidayList != null)
+            {
                 logger.LogInformation("Fetched All holiday data");
+                return holidayList;
+            }
             else
-                logger.LogError("Cannot get access to holiday collection in Db");
-            return holidayList;
+            {
+                logger.LogError("Cannot get access to Holiday collection in Db");
+                return StatusCode(500, "Internal server error");
+            } 
         }
 
         /// <summary>
@@ -48,15 +53,21 @@ namespace CASWebApi.Controllers
         public ActionResult<Holiday> GetHolidayById(string id)
         {
             logger.LogInformation("Getting holiday by given Id from HolidayController");
-            var holiday = _holidayService.GetById(id);
-
-            if (holiday == null)
+            if(id != null)
             {
-                logger.LogError("Cannot get access to holiday collection in Db");
+                var holiday = _holidayService.GetById(id);
+                if (holiday != null)
+                {
+                    logger.LogInformation("Fetched holiday data by id");
+                    return Ok(holiday);
+                }
+                else
+                    logger.LogError("Cannot get access to holiday collection in Db");
             }
-            logger.LogInformation("Fetched holiday data by id");
+            else
+                logger.LogError("Course Id is null or empty string");
+            return BadRequest(null);
 
-            return holiday;
         }
 
         /// <summary>
@@ -68,14 +79,20 @@ namespace CASWebApi.Controllers
         public ActionResult<Holiday> CreateHoliday(Holiday holiday)
         {
             logger.LogInformation("Creating a new holiday profile: " + holiday);
-            holiday.Status = true;
-            if (!(_holidayService.Create(holiday)))
+            if(holiday != null)
             {
-                logger.LogError("Cannot create a holiday, duplicated id or wrong format");
-                return NotFound(null);
+                holiday.Status = true;
+                if (_holidayService.Create(holiday))
+                {
+                    logger.LogInformation("A new holiday profile added successfully " + holiday);
+                    return CreatedAtRoute("getHolidayById", new { id = holiday.Id }, holiday);
+                }
+                else
+                    logger.LogError("Cannot create a holiday, duplicated id or wrong format");
             }
-            logger.LogInformation("A new holiday profile added successfully " + holiday);
-            return CreatedAtRoute("getHolidayById", new { id = holiday.Id }, holiday);
+            else
+                logger.LogError("Faculty object is null " + holiday);
+            return BadRequest(null);
         }
 
         /// <summary>
@@ -87,19 +104,28 @@ namespace CASWebApi.Controllers
         public IActionResult UpdateHoliday(Holiday holidayIn)
         {
             logger.LogInformation("Updating existed Holiday profile: " + holidayIn.Id);
-            var teacher = _holidayService.GetById(holidayIn.Id);
-
-            if (teacher == null)
+            if(holidayIn != null)
             {
-                logger.LogError("Holiday with Id: " + holidayIn.Id + " doesn't exist");
-                return NotFound(false);
+
+                var holiday = _holidayService.GetById(holidayIn.Id);
+
+                if (holiday != null)
+                {
+                    if (_holidayService.Update(holidayIn.Id, holidayIn))
+                    {
+                        logger.LogInformation("Given holiday profile Updated successfully");
+                        return Ok(true);
+                    }
+                    else
+                        logger.LogError("Cannot update the holiday profile: " + holidayIn.Id + " wrong format");
+                }
+                else
+                    logger.LogError("holiday with Id: " + holidayIn.Id + " doesn't exist");
+
             }
-            bool updated = _holidayService.Update(holidayIn.Id, holidayIn);
-            if (updated)
-                logger.LogInformation("Given Holiday profile Updated successfully");
             else
-                logger.LogError("Cannot update the Holiday profile: " + holidayIn.Id + " wrong format");
-            return Ok(updated);
+                logger.LogError("CourseIn objest is null");
+            return BadRequest(false);
         }
 
         /// <summary>
@@ -111,14 +137,16 @@ namespace CASWebApi.Controllers
         public IActionResult DeleteHolidayById(string id)
         {
             logger.LogInformation("Deleting existed holiday profile: " + id);
-            var holiday = _holidayService.GetById(id);
-
-            if (holiday != null && _holidayService.RemoveById(holiday.Id))
+            if(id != null)
             {
-                logger.LogInformation("holiday profile has been deleted successfully " + holiday);
-                return Ok(true);
+                var holiday = _holidayService.GetById(id);
+                if (holiday != null && _holidayService.RemoveById(holiday.Id))
+                    return Ok(true);
+                else
+                    logger.LogError("Cannot get access to faculty collection in Db");
             }
-            logger.LogError("holiday not found in Database");
+            else
+                logger.LogError("Id is not valid format or null");
             return NotFound(false);
         }
     }
